requirements.txt
django==3.1.7
djangorestframework==3.12.2
drf-yasg==1.20.0
django-filter==2.4.0
psycopg2==2.8.6